adding segment_rpc and rpc
